Harmonizing keymaps and keymap display
Mostly this is about switching so that all keymap descriptions are
lowercase by default, so that anytime you see a capital in a keymap (in
the which-key panel) you know that it's meaningful. I may ultimately
change my mind about what I want aesthetically, but we need to at least
be consistent.